Zig build run works

+++ Sources/src/Main/RPGStats.cpp
@@ -1159,8 +1159,8 @@ bool SMechUnitRPGStats::SArmor::ToAIUnits()
 }
 bool SMechUnitRPGStats::SConstraint::ToAIUnits()
 {
-	wMin = fMin >= FP_2PI ? 65535 : WORD( fMin / FP_2PI * 65535.0f );
-	wMax = fMax >= FP_2PI ? 65535 : WORD( fMax / FP_2PI * 65535.0f );
+	wMin = fMin >= FP_2PI ? 65535 : WORD( static_cast<int>( fMin / FP_2PI * 65535.0f ) );
+	wMax = fMax >= FP_2PI ? 65535 : WORD( static_cast<int>( fMax / FP_2PI * 65535.0f ) );
 	
 	return true;
 }
